Smart Climate City/requirements.txt
# Project EcoFlow - Python Dependencies

# Dashboard
streamlit>=1.28.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Time Series Forecasting
prophet>=1.1.5

# Database
psycopg2-binary>=2.9.9

# Visualization
plotly>=5.17.0

# Utilities
python-dateutil>=2.8.2
